@sjrd 2025-01-10
> If it was up to me I'd copy java's lazy initialization. If there is interest, I could prepare a PR?

That's a big no-no from a licensing point of view. The JDK source is GPL, and we are Apache 2.0. We must never copy code from the JDK, nor even copy logic from its inner workings that are not obvious from the signatures and JavaDocs.