@bbrehm 2025-01-10
Apologies, I misspoke. I wanted to say:

>If it was up to me, I'd do something similar to java's lazy initialization. If there is interest, I could prepare a PR?

That is afaiu an actual tradeoff instead of a strict improvement, so may need more discussion.