@lrytz 2025-01-16
LGTM, especially given the benchmark results (similar improvements for me).

> it pays a lot to encourage a situation where the fastpath is inlined, and the actual resizing is not inlined

Curious: did you at inline logs, do you know that the preformance improvement has to do with inlining?

In any case, this looks good to go.

We could check whether an `if` around `ensureSize` in `ArrayBuilder.addOne` also improves performance.